requirements.txt
fastapi
uvicorn
pydantic
requests
redis
python-dotenv
sqlalchemy
psycopg2-binary
passlib[argon2]
argon2-cffi
python-jose[cryptography]
fastapi-limiter
email-validator
resend